🤩🥳 Code Reorganization Bonanza!
- **next.config.ts**: Cleaned up some commented crap and kept things neat! ⚒️
  - Removal of trailing slashes in comments to reduce clutter.

- **src/app/layout.tsx** ➡️ **pages/_app.tsx**: Renamed and revamped!
  - Converted from a layout file into a proper app component. Now it’s handling routing! 🛣️
  - Replaced the `children` prop with `Component` and `pageProps` to follow Next.js conventions like a champ!

- **Created New Files**: 🆕🎉
  - **pages/index.tsx**: New homepage with a welcome message & login form! Say goodbye to boredom! 📅👋
  - **pages/profile/[did].tsx**: Dynamic profile page to display user profiles based straight off the URL! 🚀
  - **pages/profile/index.tsx**: Added a profile index page, super clean! 🌟

- **Deleted Useless Files**: 🗑️🚫
  - Cleaned up `layout.tsx`, `page.tsx`, and the `base-layout.tsx`. Less is more, baby! ✨
  - Also cleared out old CSS files that did nothing but cause confusion!

- **globals.css Renamed**: 🎨🚀
  - Slight name tweak from `src/app/globals.css` to `styles/globals.css` for better organization!

This commit cleans up the project structure, aligns with Next.js best practices, and adds some shiny new features. Users will now enjoy a smooth and intuitive experience while we’ve made the code a lot less cluttered (you’re welcome, future developers)! 😜🚀
