+++ app/pages/_app.tsx
@@ -1,32 +1,27 @@
-// src/app/layout.tsx
-"use client";
-
+// pages/_app.tsx
+import type { AppProps } from "next/app";
 import { AuthProvider } from "@/lib/auth/auth-context";
 import { ThemeProvider } from "@/lib/theme/theme-context";
 import { Inter } from "next/font/google";
 import { Logo } from "@/components/ui/logo";
-import "./globals.css";
+import "@/styles/globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
 
-export default function RootLayout({
-    children,
-}: {
-    children: React.ReactNode;
-}) {
+export default function App({ Component, pageProps }: AppProps) {
     return (
         <html lang="en">
             <body className={inter.className}>
                 <AuthProvider>
                     <ThemeProvider>
                         <div className="min-h-screen">
-                            <header className="bg-primary px-4 py-6 ">
-                                <div className="mx-auto max-w-4xl ">
+                            <header className="bg-primary px-4 py-6">
+                                <div className="mx-auto max-w-4xl">
                                     <Logo className="h-12 w-auto text-white" />
                                 </div>
                             </header>
                             <main className="mx-auto max-w-4xl px-4 py-8">
-                                {children}
+                                <Component {...pageProps} />
                             </main>
                         </div>
                     </ThemeProvider>
