+++ app/pages/profile/index.tsx
@@ -1,0 +1,5 @@
+import { ProfileContent } from "@/components/profile/profile-content";
+
+export default function ProfilePage() {
+    return <ProfileContent />;
+}
