--- app/src/app/profile/[DID]/layout.tsx
@@ -1,15 +1,0 @@
-export function generateStaticParams() {
-    return [
-        {
-            DID: "placeholder",
-        },
-    ];
-}
-
-export default function ProfileLayout({
-    children,
-}: {
-    children: React.ReactNode;
-}) {
-    return children;
-}
