--- app/src/components/layout/base-layout.tsx
@@ -1,3 +1,0 @@
-export function BaseLayout({ children }: { children: React.ReactNode }) {
-    return <>{children}</>;
-}
